Move files to base
