+++ src/base/plugins/SUBTREELoader.js
@@ -2,8 +2,8 @@
  * Structure almost identical to Cesium, also the comments and the names are kept
  * https://github.com/CesiumGS/cesium/blob/0a69f67b393ba194eefb7254600811c4b712ddc0/packages/engine/Source/Scene/Implicit3DTileContent.js
  */
-import { SubtreeTile } from '../../base/SubtreeTile.js';
-import { LoaderBase } from '../../base/loaders/LoaderBase.js';
+import { SubtreeTile } from './SubtreeTile.js';
+import { LoaderBase } from '../loaders/LoaderBase.js';
 import { readMagicBytes } from '../../utilities/readMagicBytes.js';
 import { arrayToString } from '../../utilities/arrayToString.js';
