+++ src/index.d.ts
@@ -18,8 +18,10 @@ export * from './three/renderers/CesiumIonTilesRenderer';
 // three.js plugins
 export { CesiumIonAuthPlugin } from './three/plugins/CesiumIonAuthPlugin';
 export { GoogleCloudAuthPlugin } from './three/plugins/GoogleCloudAuthPlugin';
-export { ImplicitTilingPlugin } from './three/plugins/ImplicitTilingPlugin';
 export * from './three/plugins/DebugTilesPlugin';
+
+// common plugins
+export { ImplicitTilingPlugin } from './base/plugins/ImplicitTilingPlugin';
 
 // common
 export { TilesRendererBase } from './base/TilesRendererBase';
