+++ src/index.js
@@ -24,9 +24,11 @@ export { EnvironmentControls } from './three/controls/EnvironmentControls.js';
 
 // three.js plugins
 export { CesiumIonAuthPlugin } from './three/plugins/CesiumIonAuthPlugin.js';
-export { ImplicitTilingPlugin } from './three/plugins/ImplicitTilingPlugin.js';
 export { GoogleCloudAuthPlugin } from './three/plugins/GoogleCloudAuthPlugin.js';
 export * from './three/plugins/DebugTilesPlugin.js';
+
+// common plugins
+export { ImplicitTilingPlugin } from './base/plugins/ImplicitTilingPlugin.js';
 
 // common
 export { TilesRendererBase } from './base/TilesRendererBase.js';
